requirements.txt
fastapi
uvicorn[standard]
langchain
langgraph
langchain-openai
langchain-core
requests
python-dotenv
numpy